+++ turing_stream.c
@@ -80,16 +80,16 @@ product or in the associated documentation.
 // from tivodecode.c, verbose option
 extern int o_verbose;
 
-off_t setup_turing_key(turing_state * turing, void * tivofile, read_func_t read_handler, char * mak)
+unsigned int setup_turing_key(turing_state * turing, void * tivofile, read_func_t read_handler, char * mak)
 {
     blob xml;
     SHA1_CTX context;
-    off_t mpeg_off;
+    unsigned int mpeg_off;
 
     sha1_init(&context);
     sha1_update(&context, (unsigned char *)mak, strlen(mak));
 
-    mpeg_off = (off_t)parse_tivo(tivofile, &xml, read_handler);
+    mpeg_off = parse_tivo(tivofile, &xml, read_handler);
 
     sha1_update(&context, xml.data, xml.size);
     sha1_final(turing->turingkey, &context);
